mnist_ml/requirements.txt
celery==5.3.6
kombu==5.3.5
redis==5.0.2
scikit_learn==1.3.0
SQLAlchemy==1.4.39
torch==1.8.0
eventlet
python-dotenv
psycopg2-binary
torchvision